Refactor shared trash finesse logic into common connection helper.
